tests/wpt/web-platform-tests/dom/nodes/Document-contentType/contentType/contenttype_datauri_02.html

<!DOCTYPE html>
<title>Data URI document.contentType === 'text/html' when data URI MIME type is set</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var iframe = document.createElement('iframe');
  iframe.addEventListener('load', this.step_func_done(function() {
    assert_equals(iframe.contentDocument.contentType, "text/html");
  }), false);
  iframe.src = "data:text/html;charset=utf-8,<!DOCTYPE html>";
  document.body.appendChild(iframe);
});
</script>
